# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Eduator Clean Backend API
  version: 0.4.0
  description: >-
    Clean backend API contract for local PostgreSQL + JWT + HTTP API keys (baseline **0.4.0**).
    Includes OpenRouter multi-model AI (migration `018_ai_provider_openrouter.sql`),
    pgvector-backed RAG (migration `015_pgvector_document_chunks.sql`), user HTTP API keys,
    usage analytics (`api_access_log` after migration `006_api_access_log.sql`),
    and first-party traffic exclusion via `X-Eduator-Client` (see docs).
servers:
  - url: http://localhost:4000
    description: Local backend
...
paths:
...
  /v1/admin/ai-providers:
    get:
      tags: [Admin]
      summary: Get OpenRouter credential status, workload policies, and model catalog
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: AI provider admin overview
        '403':
          description: Admin JWT required
  /v1/admin/ai-providers/credential:
    put:
      tags: [Admin]
      summary: Save encrypted platform OpenRouter API key
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Credential saved
    delete:
      tags: [Admin]
      summary: Delete platform OpenRouter API key
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Credential deleted
  /v1/admin/ai-providers/credential/test:
    post:
      tags: [Admin]
      summary: Test the configured OpenRouter credential
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Test result
  /v1/admin/ai-providers/catalog/sync:
    post:
      tags: [Admin]
      summary: Sync OpenRouter model catalog into the database
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Catalog synced
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...